pyproject.toml
[project]
name = "harmful-rag-based-bots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-aiplatform>=1.128.0",
    "google-cloud-discoveryengine>=0.15.0",
    "google-cloud-storage>=3.6.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
